Jsonnet: replace null with {} (#780)
Converting `null` or `{}` to YAML both result in `null`. But null in Jsonnnet can trip someone up that expects the value to be there. For instance when extending tempo_config they might do something like this:

```
  tempo_config+:: {
    distributor+: {
      // ...
    }
  }
```

Using +: is generally preferred because it doesn't overwrite previous values, but would not work if tempo_config.distributor is null.