Allow compactor disablement (#1850)
* Allow compactor disablement

Without this change, there isn't a good way to disable compaction when
running in single binary or scalable single binary mode.  Normally, when
running in microservices mode, operators can simply scale down the
compactors to 0 to effectively disable compaction, though this will have
other consequences, since the compactors are doing a little more than
compaction.

Here we include a config option to disable compaction at the compactor
level.  This allows the compactor module to run, but disable compaction,
allowing blocks to remain in the backend for debug or investigation
purposes.

* Update docs

* Update changelog

* Fix config path reference