[generator] Insert leading zero value for new counter series (#2068)
* [generator] insert leading 0 on new counter series

New counters are created with null values, which causes Prometheus to
ignore the first value to avoid the comparison of null -> x.  Here we
allow for first values on new series to be more useful by appending
first a zero to the series, followed by a value.  This will allow
Prometheus queries to use the first value of 0 -> x rather than null ->
x.

* [generator] udpate generator tests for zero-leading on new series
* Replace firstSeries mutex with atomic.Bool
* Offset the initial counter value by 1 second
* Update changelog
* Move registerSeenSeries() where we expect it to be needed
---------

Signed-off-by: Zach Leslie <zach.leslie@grafana.com>