Fix GetTrace() in tempo-query (#5864)
* Fix GetTrace() in tempo-query

Upgrading to grpc 1.38.0 broke compatibility with gogoproto.customtype,
and the trace_id field in GetTraceRequest of storage.pb.go uses customtype.

This issue only occurs when navigating to a trace in Jaeger UI and
then hitting refresh (this triggers a GetTrace() request).
Searching traces in Jaeger UI and opening them worked fine before this
change as well, because this issued a FindTraces() request, which doesn't
use a customtype.

* add changelog entry

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>

---------

Signed-off-by: Andreas Gerstmayr <agerstmayr@redhat.com>