[Metrics Generator] Add support for entity limiting (#5788)
* [Generator] Add entity limiter

This PR adds a new type of limiter to the metrics generator. It is
designed to limit on "entities" rather than series. In this case an
entity is a single labelset across multiple metrics, excluding external
labels. In effect, it allows the limiter to always generate the full set
of data for a given entity, rather than limiting randomly once the
series limit is triggered.

Importantly, this limit is applied at collection time, so it's still
necessary to set a series limit in order to avoid generator OOM kills.
In practice, the series limit should likely be some constant multiple of
the entity limit. e.g. 20x. This allows the system to primarily limit on
entities while still enforcing a ceiling on memory consumption.

* fix lint

* Fix legacy overrides

* fix config_test

* Add metrics for limits and demand

* Separate demand ticker

* Fix rebase

* Use clear and seq.Iters

* Use separate `deleteByHash` instead of overloading deleteFunc

* reduce diff on naming

* Remove unnecessary diff

* Remove unnecessary diff 2

* Fix flaky test

* Extract series limiter and simplify limiter interface

* Satisfy linter

* More ci

* metric parity with series:

* Move cardinality back

* Fix references

* Remove delete by hash

* Add missing assertion

* Address PR feedback

* Fix unit tests

* Fix entity demand + add OnUpdate callback